+++ src/main/java/com/nhnacademy/illuwa/review/service/ReviewService.java
@@ -3,9 +3,7 @@ package com.nhnacademy.illuwa.review.service;
 import com.nhnacademy.illuwa.review.client.ReviewServiceClient;
 import com.nhnacademy.illuwa.review.dto.ReviewRequest;
 import com.nhnacademy.illuwa.review.dto.ReviewResponse;
-import com.nhnacademy.illuwa.user.member.client.MemberServiceClient;
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
+import com.nhnacademy.illuwa.member.client.MemberServiceClient;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
