+++ src/main/resources/static/css/page/book/booklist.css
